app/requirements.txt
python-logging-loki

